+++ synapse/src/test/resources/report/test-project-report.html
@@ -265,32 +265,31 @@
   <div class="summary-container">
     <h2>Migration Coverage Overview</h2>
     <div class="metrics" style="flex-direction: column; align-items: center; width: 100%;">
-
-      <!-- Overall Confidence -->
       <div class="metric">
-        <div class="metric-left">
-          <span class="metric-value">82%</span>
-          <span class="metric-label">Overall Confidence</span>
-          <div class="coverage-indicator">
-            <div class="coverage-bar" style="width: 82%; background-color: #4CAF50;"></div>
-          </div>
-          <div class="status-badge-container">
-            <span class="status-badge status-high">HIGH CONFIDENCE</span>
-          </div>
-        </div>
-        <div class="metric-right">
-          <div class="coverage-breakdown">
-            <div style="display: flex; justify-content: space-between; margin-bottom: 4px;">
-              <span class="breakdown-label">Total Code Lines:</span>
-              <span class="breakdown-value">173</span>
-            </div>
-            <div style="display: flex; justify-content: space-between; margin-bottom: 4px;">
-              <span class="breakdown-label">Mediators:</span>
-              <span class="breakdown-value">10</span>
-            </div>
-          </div>
-        </div>
+  <div class="metric-left">
+    <span class="metric-value">82%</span>
+    <span class="metric-label">Overall Confidence</span>
+    <div class="coverage-indicator">
+      <div class="coverage-bar" style="width: 82%; background-color: #4CAF50;"></div>
+    </div>
+    <div class="status-badge-container">
+      <span class="status-badge status-high">HIGH CONFIDENCE</span>
+    </div>
+  </div>
+  <div class="metric-right">
+    <div class="coverage-breakdown">
+      <div style="display: flex; justify-content: space-between; margin-bottom: 4px;">
+        <span class="breakdown-label">Total Code Lines:</span>
+        <span class="breakdown-value">173</span>
       </div>
+      <div style="display: flex; justify-content: space-between; margin-bottom: 4px;">
+        <span class="breakdown-label">Mediators:</span>
+        <span class="breakdown-value">10</span>
+      </div>
+    </div>
+  </div>
+</div>
+
     </div>
   </div>
   <div class="summary-container">
@@ -347,6 +346,7 @@
 
       </ul>
     </div>
+
   </div>
   <div class="summary-container">
     <h2>Mediator wise breakdown</h2>
@@ -416,5 +416,5 @@
   </table>
   </div>
 </div>
-<footer><p>Report generated on: 7/24/2025, 1:57:25 PM</p></footer>
+<footer><p>Report generated on: 7/24/2025, 3:21:45 PM</p></footer>
 </body></html>
